+++ src/main/java/ru/efimov/DiplomFirst/controller/ExitController.java
@@ -4,9 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import ru.efimov.DiplomFirst.entity.Enter;
 import ru.efimov.DiplomFirst.entity.Exit;
+import ru.efimov.DiplomFirst.entity.Student;
 import ru.efimov.DiplomFirst.repository.EnterRepository;
 import ru.efimov.DiplomFirst.repository.ExitRepository;
 import ru.efimov.DiplomFirst.repository.StudentRepository;
@@ -31,6 +33,13 @@ public class ExitController {
         if (!studentRepository.existsById(studentId)) {
             throw new ResourceNotFoundException("Not found Student with id = " + studentId);
         }
+        Student student1 = studentRepository.findById(studentId)
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + studentId));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
 
         List<Exit> exits = exitRepository.findByStudentId(studentId);
         return new ResponseEntity<>(exits, HttpStatus.OK);
@@ -41,12 +50,27 @@ public class ExitController {
         Exit exit = exitRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Not found Exit with id = " + id));
 
+        Student student = studentRepository.findById(exit.getStudent().getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + exit.getStudent().getId()));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         return new ResponseEntity<>(exit, HttpStatus.OK);
     }
 
     @PostMapping("/students/{studentId}/exits")
     public ResponseEntity<Exit> createExit(@PathVariable(value = "studentId") Long studentId,
                                              @RequestBody Exit exitRequest) {
+        Student student1 = studentRepository.findById(studentId)
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + studentId));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+
         Exit exit = studentRepository.findById(studentId).map(student -> {
             exitRequest.setStudent(student);
             return exitRepository.save(exitRequest);
@@ -60,6 +84,13 @@ public class ExitController {
         Exit exit = exitRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("ExitId " + id + "not found"));
 
+        Student student = studentRepository.findById(exit.getStudent().getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + exit.getStudent().getId()));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         exit.setDate_of_exit(exitRequest.getDate_of_exit());
 
         return new ResponseEntity<>(exitRepository.save(exit), HttpStatus.OK);
@@ -67,6 +98,16 @@ public class ExitController {
 
     @DeleteMapping("/exits/{id}")
     public ResponseEntity<HttpStatus> deleteExit(@PathVariable("id") long id) {
+        Exit exit = exitRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("ExitId " + id + "not found"));
+
+        Student student = studentRepository.findById(exit.getStudent().getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + exit.getStudent().getId()));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         exitRepository.deleteById(id);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -77,6 +118,14 @@ public class ExitController {
         if (!studentRepository.existsById(studentId)) {
             throw new ResourceNotFoundException("Not found Student with id = " + studentId);
         }
+        Student student1 = studentRepository.findById(studentId)
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + studentId));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+
 
         exitRepository.deleteByStudentId(studentId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
