+++ src/main/java/ru/efimov/DiplomFirst/controller/OpenMaterialController.java
@@ -4,11 +4,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
-import ru.efimov.DiplomFirst.entity.Enter;
-import ru.efimov.DiplomFirst.entity.Material;
-import ru.efimov.DiplomFirst.entity.OpenMaterial;
-import ru.efimov.DiplomFirst.entity.Student;
+import ru.efimov.DiplomFirst.entity.*;
 import ru.efimov.DiplomFirst.repository.EnterRepository;
 import ru.efimov.DiplomFirst.repository.MaterialRepository;
 import ru.efimov.DiplomFirst.repository.OpenMaterialRepository;
@@ -40,7 +38,18 @@ public class OpenMaterialController {
         }
 
         List<OpenMaterial> openMaterials = openMaterialRepository.findByMaterialId(materialId);
-        return new ResponseEntity<>(openMaterials, HttpStatus.OK);
+
+        List<OpenMaterial> openMaterialsCurStudent = new ArrayList<>();
+        for (OpenMaterial o1 : openMaterials){
+            Student student1 = studentRepository.findById(o1.getStudent().getId())
+                    .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + o1.getStudent().getId()));
+
+            if (SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+                openMaterialsCurStudent.add(o1);
+            }
+        }
+
+        return new ResponseEntity<>(openMaterialsCurStudent, HttpStatus.OK);
     }
 
     @GetMapping("/students/{studentId}/openMaterials")
@@ -48,6 +57,13 @@ public class OpenMaterialController {
         if (!studentRepository.existsById(studentId)) {
             throw new ResourceNotFoundException("Not found Student with id = " + studentId);
         }
+        Student student1 = studentRepository.findById(studentId)
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + studentId));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
 
         List<OpenMaterial> openMaterials = openMaterialRepository.findByStudentId(studentId);
         return new ResponseEntity<>(openMaterials, HttpStatus.OK);
@@ -58,6 +74,13 @@ public class OpenMaterialController {
         OpenMaterial openMaterial = openMaterialRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Not found OpenMaterial with id = " + id));
 
+        Student student1 = studentRepository.findById(openMaterial.getStudent().getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + openMaterial.getStudent().getId()));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         return new ResponseEntity<>(openMaterial, HttpStatus.OK);
     }
 
@@ -65,6 +88,13 @@ public class OpenMaterialController {
     public ResponseEntity<OpenMaterial> createOpenMaterial(@PathVariable(value = "studentId") Long studentId,
                                                            @PathVariable(value = "materialId") Long materialId,
                                              @RequestBody OpenMaterial openMaterialRequest) {
+        Student student1 = studentRepository.findById(studentId)
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + studentId));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         Optional<Student> optionalStudent = studentRepository.findById(studentId);
         Optional<Material> optionalMaterial = materialRepository.findById(materialId);
         Student student = null;
@@ -95,6 +125,13 @@ public class OpenMaterialController {
         OpenMaterial openMaterial = openMaterialRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("OpenMaterialId " + id + "not found"));
 
+        Student student1 = studentRepository.findById(openMaterial.getStudent().getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + openMaterial.getStudent().getId()));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         openMaterial.setDate_of_open(openMaterialRequest.getDate_of_open());
 
         return new ResponseEntity<>(openMaterialRepository.save(openMaterial), HttpStatus.OK);
@@ -102,6 +139,16 @@ public class OpenMaterialController {
 
     @DeleteMapping("/openMaterials/{id}")
     public ResponseEntity<HttpStatus> deleteOpenMaterial(@PathVariable("id") long id) {
+        OpenMaterial openMaterial = openMaterialRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("OpenMaterialId " + id + "not found"));
+
+        Student student1 = studentRepository.findById(openMaterial.getStudent().getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + openMaterial.getStudent().getId()));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
         openMaterialRepository.deleteById(id);
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -112,6 +159,13 @@ public class OpenMaterialController {
         if (!studentRepository.existsById(studentId)) {
             throw new ResourceNotFoundException("Not found Student with id = " + studentId);
         }
+        Student student1 = studentRepository.findById(studentId)
+                .orElseThrow(() -> new ResourceNotFoundException("Not found Student with id = " + studentId));
+
+        if (!SecurityContextHolder.getContext().getAuthentication().getPrincipal().equals(student1.getUsername())){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
 
         openMaterialRepository.deleteByStudentId(studentId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
